0x1A-hash_tables - 0 add edge case size 0 return hash table

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *new = NULL;
 
-	if (size < 0)
+	if ((int)size < 0)
 		return (NULL);
 
 	new = malloc(sizeof(hash_table_t)); /**/
